+++ src/views/FitnessGoalsList.vue
@@ -3,6 +3,7 @@ import FitnessGoalGrid from "../components/FitnessGoalGrid.vue";
 import PageHeader from "../components/PageHeader.vue";
 import FitnessGoalsDataService from "../services/FitnessGoalsDataService";
 import WidgetAddButton from "../components/WidgetAddButton.vue";
+import Swal from "sweetalert2";
 </script>
 
 <script>
@@ -45,15 +46,28 @@ export default {
     },
 
     deleteFitnessGoal(itemId, index) {
-      FitnessGoalsDataService.delete(itemId)
-        .then((response) => {
-          //console.log(response.data);
-          this.items.splice(index, 1);
-          this.page.message = this.items.length == 0 ? "No records found." : "";
-        })
-        .catch((error) => {
-          console.log(error);
-        });
+      Swal.fire({
+        title: "DELECTION CONFIRMATION!",
+        text: "Are you sure that you want to delete the item?",
+        icon: "warning",
+        confirmButtonText: "Yes, delete it.",
+        cancelButtonText: "No, cancel.",
+        showCancelButton: true,
+        showCloseButton: true,
+      }).then((result) => {
+        if (result.isConfirmed) {
+          FitnessGoalsDataService.delete(itemId)
+            .then((response) => {
+              //console.log(response.data);
+              this.items.splice(index, 1);
+              this.page.message =
+                this.items.length == 0 ? "No records found." : "";
+            })
+            .catch((error) => {
+              console.log(error);
+            });
+        }
+      });
     },
   },
 };
